moved pages to components file, added some styling for the single product page and contact form page, fixed bug on back end
