+++ client/src/App.jsx
@@ -6,7 +6,7 @@ import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 
 // Importing pages
 import Header from "./components/Header"
-import ContactForm from "./pages/Contact";
+import ContactForm from "./components/Contact";
 import Footer from "./components/Footer";
 import Home from "./pages/Home";
 
@@ -28,9 +28,9 @@ function App() {
                 <Routes>
                     <Route path="/" element={<Home />} />
                     {/* <Route path="/products" element={<Products />} /> */}
-                    {/* <Route path="/contacts" element={<ContactForm />} />
-                    <Route path="/profile" element={<Profile />} />
-                    <Route path="/cart" element={<Cart />} /> */}
+                    <Route path="/contact" element={<ContactForm />} />
+                    {/* <Route path="/profile" element={<Profile />} /> */}
+                    {/* <Route path="/cart" element={<Cart />} /> */}
                 </Routes>
                 <Footer />
             </Router>
