+++ client/src/components/Contact.jsx
@@ -21,12 +21,16 @@ const ContactForm = () => {
     }
 
     return (
+        <>
+        <h2 className="text-center text-3xl">Questions?</h2>
+        <p className="text-center m-3">If you'd like a customer service representative to contact you for a specific order, please fill out the form below and we'll be in touch soon!</p>
         <form
+            className="m-5 grid place-items-center"
             action={FORM_ENDPOINT}
             onSubmit={handleSubmit}
             method="POST"
             target="_blank">
-            <div className="mb-3 pt-0">
+            <div className="mb-3 w-1/2">
                 <input
                     type="text"
                     placeholder="Your Name"
@@ -35,7 +39,7 @@ const ContactForm = () => {
                     required
                 />
             </div>
-            <div className="mb-3 pt-0">
+            <div className="mb-3 pt-0 w-1/2">
                 <input
                     type="email"
                     placeholder="Email"
@@ -44,7 +48,7 @@ const ContactForm = () => {
                     required
                 />
             </div>
-            <div className="mb-3 pt-0">
+            <div className="mb-3 pt-0 w-1/2">
                 <textarea
                     placeholder="Your message"
                     name="message"
@@ -53,13 +57,14 @@ const ContactForm = () => {
                 />
             </div>
             <div className="mb-3 pt-0">
-                <button className="bg-blue-500 text-white active:bg-blue-600 font-bold uppercase text-sm px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
+                <button className="bg-neutral-500 text-white active:bg-neutral-700 font-bold uppercase text-sm px-6 py-3 rounded shadow hover:shadow-lg outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
                     type="submit"
                 >
                     Send Message
                 </button>
             </div>
         </form>
+        </>
     );
 };
