+++ client/src/index.css
@@ -8,8 +8,8 @@
     /* line-height: 1.5; */
     font-weight: 400;
     color-scheme: light dark;
-    color: #11402f;
-    background-color: #f2f1e9;
+    color: #f2f1e9;
+    background-color: #11402f;
   
     font-synthesis: none;
     text-rendering: optimizeLegibility;
@@ -18,7 +18,7 @@
     -webkit-text-size-adjust: 100%;
   }
   
-}
+};
