+++ server/schemas/resolvers.js
@@ -5,9 +5,6 @@ const resolvers = {
         products: async () => {
             return await Product.find({});
         }, 
-        product: async (parent, { productId }) => {
-            return await Product.findOne({_id: productId });
-        }, 
         users: async () => {
             return await User.find({});
         }
@@ -15,3 +12,7 @@ const resolvers = {
 };
 
 module.exports = resolvers;
+
+// [product: async (parent, { productId }) => {
+//     return await Product.findOne({_id: productId });
+// }, ]
